+++ CHANGELOG.md
@@ -36,6 +36,9 @@
   * Fixed a few on-hit effects still being triggered while effect is active:
     * Some Slow effects, Tar slow effect (ex. from Dunestrider), Void Reaver Nullify, Perfected Elites Cripple, Merc Expose
 
+* UI:
+  * Fixed Next Effect display still being visible when effect activation is disabled
+
 * Audio:
   * Fixed effect activation sound sometimes not playing for effects that move or change the player character